style: restructure the styles
