+++ src/components/Header/Header.style.tsx
@@ -1,0 +1,12 @@
+import Image from 'next/image';
+import styled from 'styled-components';
+
+export const WrapperImage = styled.div``;
+
+export const Info = styled.div`
+  margin-top: 60px;
+`;
+export const HeaderImage = styled(Image)`
+  width: 100%;
+  height: auto;
+`;
