+++ src/components/common/Article2.tsx
@@ -1,12 +1,12 @@
-import { AContainer, Wrapper } from '@/styles/sharedstyles';
+import * as SharedStyles from '@/styles/sharedstyles';
 import { AProps } from '@/utils/types';
 import Description from './Description';
 import Titles from './Titles';
 
 const Article2 = (props: AProps) => {
   return (
-    <AContainer>
-      <Wrapper>
+    <SharedStyles.AContainer>
+      <SharedStyles.Wrapper>
         {props.data.map((element, i) => (
           <Titles
             fontSize={element.fontSize}
@@ -15,7 +15,7 @@ const Article2 = (props: AProps) => {
             key={i}
           />
         ))}
-      </Wrapper>
+      </SharedStyles.Wrapper>
       <div>
         {props.data.map((element, i) => (
           <div key={i}>
@@ -23,7 +23,7 @@ const Article2 = (props: AProps) => {
           </div>
         ))}
       </div>
-    </AContainer>
+    </SharedStyles.AContainer>
   );
 };
