+++ src/components/common/Description.tsx
@@ -1,8 +1,8 @@
-import { Text } from '@/styles/sharedstyles';
+import * as SharedStyles from '@/styles/sharedstyles';
 interface IDescriptionProps {
   description: string;
 }
 const Description: React.FC<IDescriptionProps> = (props) => {
-  return <Text>{props.description}</Text>;
+  return <SharedStyles.Text>{props.description}</SharedStyles.Text>;
 };
 export default Description;
