+++ src/components/common/navbar/Navbar.tsx
@@ -1,90 +1,8 @@
 import React, { useState } from 'react';
-import Image from 'next/image';
-import { IoMenuSharp, IoCloseSharp } from 'react-icons/io5';
-import styled from 'styled-components';
+import Link from 'next/link';
 import List from '@/components/common/navbar/List';
 // eslint-disable-next-line import/order
-import { ToggleProps } from '@/utils/types';
-
-const Nav = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background-color: ${(props) => props.theme.colors.blackLight};
-  color: ${(props) => props.theme.colors.white};
-  padding: 10px;
-  @media (max-width: 939px) {
-    justify-content: space-between;
-    width: 100%;
-    position: fixed;
-  }
-`;
-
-const ListItems = styled.ul<ToggleProps>`
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  @media (max-width: 939px) {
-    display: ${(props) => (props.toggle === false ? 'none' : 'block')};
-    flex-direction: column;
-    margin-top: 128px;
-    height: 100vh;
-}
-}
-  }
-`;
-
-const WrapperLogo = styled.div`
-  display: flex;
-  justify-content: center;
-`;
-
-const Drawer = styled.div`
-  display: none;
-  @media (max-width: 939px) {
-    display: flex;
-    justify-content: space-between;
-  }
-`;
-
-const Wrapper = styled.div<ToggleProps>`
-  background-color: ${(props) => props.theme.colors.blackLight};
-  @media (max-width: 939px) {
-    position: fixed;
-    width: 100%;
-  }
-  @media (min-width: 13010px) {
-    display: flex;
-    justify-content: space-around;
-    width: 100%;
-  }
-`;
-const MenuSharp = styled(IoMenuSharp)`
-  font-size: 6em;
-  @media (max-width: 939px) {
-    font-size: 3em;
-  }
-`;
-const CloseSharp = styled(IoCloseSharp)`
-  font-size: 6em;
-  @media (max-width: 939px) {
-    font-size: 3em;
-  }
-`;
-const Logo = styled(Image)``;
-
-const Button = styled.button`
-  border: 1px solid #fff;
-  background: ${(props) => props.theme.colors.blackLight};
-  border-radius: 5px;
-  color: ${(props) => props.theme.colors.white};
-  padding: 10px 24px;
-  @media (max-width: 960px) {
-    margin-left: 20px;
-    margin-top: 18px;
-  }
-`;
+import * as S from './Navbar.style';
 
 const Navbar: React.FC = () => {
   const [toggle, setToggle] = useState(false);
@@ -94,24 +12,24 @@ const Navbar: React.FC = () => {
   };
 
   return (
-    <Wrapper toggle={toggle}>
-      <Nav>
-        <WrapperLogo>
-          <Logo src={'/logo.svg'} alt="Logo" width={100} height={100} />
-        </WrapperLogo>
-        <Drawer>
-          {toggle === false && <MenuSharp onClick={handleClick} />}
-          {toggle === true && <CloseSharp onClick={handleClick} />}
-        </Drawer>
-      </Nav>
+    <S.Wrapper toggle={toggle}>
+      <S.Nav>
+        <S.WrapperLogo>
+          <S.Logo src={'/logo.svg'} alt="Logo" width={100} height={100} />
+        </S.WrapperLogo>
+        <S.Drawer>
+          {toggle === false && <S.MenuSharp onClick={handleClick} />}
+          {toggle === true && <S.CloseSharp onClick={handleClick} />}
+        </S.Drawer>
+      </S.Nav>
 
-      <ListItems toggle={toggle}>
+      <S.ListItems toggle={toggle}>
         <List />
-        <Button>
-          <a href={'/login'}>LOGIN</a>
-        </Button>
-      </ListItems>
-    </Wrapper>
+        <S.Button>
+          <Link href={'/login'}>LOGIN</Link>
+        </S.Button>
+      </S.ListItems>
+    </S.Wrapper>
   );
 };
 export default Navbar;
