+++ src/components/our_menu/OurMenu.style.tsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const Container = styled.div`
+  display: flex;
+  align-items: center;
+  flex-flow: column nowrap;
+`;
