+++ src/components/what_we_are_serving/Serving.style.tsx
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const Container = styled.div`
+  width: 100%;
+  background: ${({ theme }) => theme.colors.black};
+  @media (max-width: 430px) {
+    background: none;
+  }
+`;
