+++ src/pages/about.tsx
@@ -1,16 +1,9 @@
 import Article1 from '@/components/common/Article1';
 import CustomImage from '@/components/common/CustomImage';
-import Layout from '@/components/common/Layout';
+import Layout from '@/components/common/Layout/Layout';
 import Reputation from '@/components/common/Reputation';
 import Titles from '@/components/common/Titles';
-import {
-  FDescription,
-  FImg,
-  FSubTitle,
-  FTitle,
-  FWrapper,
-  Features,
-} from '@/components/know_more_about_us/KnowMoreAboutUs';
+import * as S from '@/components/know_more_about_us/KnowMoreAboutUs.style';
 import Timeline from '@/components/know_more_about_us/Timeline';
 import Map from '@/components/Map/Map';
 import { Section, Section2, Wrapper } from '@/styles/sharedstyles';
@@ -25,11 +18,11 @@ const About = () => {
         <Section>
           <Wrapper>
             <Article1 data={data['two']} />
-            <Features>
+            <S.Features>
               {data['timeLine1'].map((element, i) => (
-                <FWrapper key={i}>
-                  <FTitle>
-                    <FImg
+                <S.FWrapper key={i}>
+                  <S.FTitle>
+                    <S.FImg
                       src={element.image}
                       alt="feature-icon"
                       width={20}
@@ -37,12 +30,12 @@ const About = () => {
                       sizes="100vh"
                     />
                     {element.title}
-                  </FTitle>
-                  <FSubTitle>{element.subTitle}</FSubTitle>
-                  <FDescription>{element.description}</FDescription>
-                </FWrapper>
+                  </S.FTitle>
+                  <S.FSubTitle>{element.subTitle}</S.FSubTitle>
+                  <S.FDescription>{element.description}</S.FDescription>
+                </S.FWrapper>
               ))}
-            </Features>
+            </S.Features>
           </Wrapper>
           <CustomImage data={data['featuresImage2']} />
         </Section>
