+++ src/pages/contact.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import Article1 from '@/components/common/Article1';
 import CustomImage from '@/components/common/CustomImage';
-import Layout from '@/components/common/Layout';
-import ContactForm from '@/components/ContactForm/ContactForm';
+import Layout from '@/components/common/Layout/Layout';
+import ContactForm from '@/components/Form/ContactForm';
 import Map from '@/components/Map/Map';
-import { Section, Wrapper } from '@/styles/sharedstyles';
+import * as SharedStyles from '@/styles/sharedstyles';
 import { data } from '@/utils/data';
 import IndexHead from './IndexHead';
 
@@ -15,13 +15,13 @@ const Contact = (props: Props) => {
     <div>
       <IndexHead />
       <Layout>
-        <Section>
-          <Wrapper>
+        <SharedStyles.Section>
+          <SharedStyles.Wrapper>
             <Article1 data={data['eight']} />
             <ContactForm />
-          </Wrapper>
+          </SharedStyles.Wrapper>
           <CustomImage data={data['featuresImage4']} />
-        </Section>
+        </SharedStyles.Section>
         <Map />
       </Layout>
     </div>
