+++ src/pages/story.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import Article1 from '@/components/common/Article1';
 import CustomImage from '@/components/common/CustomImage';
-import Layout from '@/components/common/Layout';
+import Layout from '@/components/common/Layout/Layout';
 import Map from '@/components/Map/Map';
-import { Section, Wrapper } from '@/styles/sharedstyles';
+import * as SharedStyles from '@/styles/sharedstyles';
 import { data } from '@/utils/data';
 import IndexHead from './IndexHead';
 
@@ -14,12 +14,12 @@ const Story = (props: Props) => {
     <div>
       <IndexHead />
       <Layout>
-        <Section>
-          <Wrapper>
+        <SharedStyles.Section>
+          <SharedStyles.Wrapper>
             <Article1 data={data['seven']} />
-          </Wrapper>
+          </SharedStyles.Wrapper>
           <CustomImage data={data['featuresImage3']} />
-        </Section>
+        </SharedStyles.Section>
         <Map />
       </Layout>
     </div>
